Sema: Fix historical quirk with @escaping diagnostics
Saying "implicitly non-escaping because it was declared @autoclosure"
does not make sense. Parameters of function type are now always
non-escaping by default, whether or not they are @autoclosure.

This note is a holdover from the Swift 2 days, where escaping was
the default, and @autoclosure implied @noescape.

Now, always emit the basic "implicitly non-escaping" note.