[sil] Change two SILInstruction::removeFromParent => SILInstruction::eraseFromParent.
removeFromParent should only be used when one is moving a SILInstruction from
one block to another or in preparation for deleting an instruction. In this
case, we are removing these two SILInstructions from their parent block and
leaking the instruction. The right thing to do is to use eraseFromParent.

<rdar://problem/31276565>

(cherry picked from commit 5fa67e0adbc9a12daf991e738494dfc20acaf9be)