SILGen: Fix super_method calls from class method returning Self
Calls of superclass methods across module boundaries appear to use
the super_method instruction, even if the module was not built
with resilience.

This instruction was being created with a value having a dynamic
Self type, which tripped up the SIL verifier; with the verifier
disabled we would crash in IRGen.

Fixes <rdar://problem/32667187>.