Sema: Fix bugs when applying solutions containg type(of:)
When applying a solution to an ApplyExpr, we have to coerce the
argument to the input type of the function. This is because the
solution succeeds if the argument is convertible to the
function's input, not just if they are equal.

We were forgetting to do that for the special case of a type(of:),
which has its own semantics.

Fixes <rdar://problem/32435723> and
<https://bugs.swift.org/browse/SR-5168>.