[ClangImporter] Honor the 'DC' parameter for importStringLiteral.
We don't have any way to import a macro into a different DeclContext
anyway, but even then this saves a bit of unnecessary work.

Groundwork for rdar://problem/32199805, which is that Clang has
removed a bunch of facilities for going from a SourceLocation or
MacroInfo to a Module in favor of things like clang::ModuleMacro.