Collate Codable diagnostics for correct ordering
Codable failures produced diagnostics and notes explaining what failed
during decode. These diagnostics were being output as they were
generated, which was suboptimal because it meant that NOTEs could be
produced before the error they were related to.

This was bad for consumers like Xcode, which associate NOTEs with the
last error or warning diagnostic.

This collates these notes using a diagnostic transaction and only
outputs the errors if something goes wrong.