[capture-promotion] Be sure to copy_value a struct_extracted value from a promoted capture even though the value is borrowed.
The issue here is that we know that the given value is already borrowed, so
technically, we do not need a copy. The problem is that there is still
potentially a destroy of the value that will be cloned by the closure cloner.
This results in use-after-frees.

This commit fixes the problem by inserting the relevant copy_value and changing
the borrowed value into a +1 copy.

This is a bug that would have been caught by the Ownership verifier since a
borrowed value can not be destroyed (so the verifier would have tripped).

rdar://32625475