Use a cast instead of trying to rely on higher level APIs.
I missed this problem due to a missing test case in initializers.swift. In a
subsequent commit, I am going to add the rest of the missing cases with some
FileCheck tests for some of them. This is to ensure that the commit is easy to
review.

rdar://32539006
(cherry picked from commit cd55f466260b41b95cb0d113f9d7c6c86f6b0304)