[SILGen] Update the maximum value in a find-index-with-largest-value loop.
This is a loop that attempts to find the index of something with the longest
prefix, but wasn't actually saving the new longest length when finding something
with a longer length, meaning the loop always chose the last index. The update was
accidentally lost in f40e9ef0aec626fed0214ae372e1b55d5247481d.